requirement.txt
flask
faiss-cpu
torch
transformers
sentence-transformers
python-dotenv
huggingface-hub
